Add option for advanced navigation to existing dropdown, add table in DB
The new table holds information about dependend questions and the
corresponding answer. A question can have multiple dependencies. In
dependlogic a value can specify wether the dependchoice is handled as
false or true condition (go to this question if the answer IS NOT this
choice-id / if the anser IS this choice-id).